refactor: feature flag hooks
